Set `--pull never` when running Docker containers we built
Also adds `--no-cache` to `orchestration/` when building the runtime for
consistency with how it's ran in the test environment to ensure the procedure
works.

Ensures Docker, for whatever reason, never attempts to pull a container (of
arbitrary contents) we expect to have locally built.

Also replaces `--progress plain` (`buildx`) with `--quiet` (`build`). The
The greater compatibility is preferred over the additional logging.