Use `busybox` as the image to further export the WASM with
The existing `FROM scratch` didn't work as it inherited the builder's platform
(fixed to `linux/amd64`). Once built on `linux/amd64`, we want any host to be
able to export the artifact.

For the choice of `busybox:musl`, it's equivalent to the choice of `alpine`
(which ships `busybox` dynamically-linked to `musl`), solely even further
minimized. We also don't require the exporting image be reproducible, solely
the WASM itself.