Remove `mini-serai`
`loom` is a really interesting tool, and `mini-serai` was critical when it was
written as a proof of concept for a very complicated issue. We now have
[comprehensive documentation on multisig rotation](
  https://github.com/serai-dex/serai/blob/next-polkadot-sdk/spec/processor/Multisig%20Rotation.md
) however, and this has fallen by the wayside. It also isn't something I'd say
is actively worth moving forward with. Instead, I'd say it's something to be
mindful of and always look for a potential way to incorporate in the future.

This closes https://github.com/serai-dex/serai/issues/378 as "won't fix".