When emulating the runtime builder in CI, only perform a single build
This is only justifiable due to it being cross-checked against other instances.
If this was done on its own, it'd invalidate the test. It isn't declared as a
feature so we are able to use `--all-features` without invalidating the test,
and while we could have the proper amount of runs be a feature, that'd prevent
using `-all-features` in CI, hence the use of a `cfg`.