Version `Signature`
https://github.com/serai-dex/serai/issues/639 is relevant.

Currently, we can get away with using `RistrettoSignature` everywhere. In the
future, we may want to have transaction signatures be ML-DSA yet we'd need a
distinct solution for setting keys (even if just `BoundedVec<MlDsaSignature>`)
and yet another solution for publishing `Batch`es/`SlashReport`s. The current
usages of `Signature` are documented to potentially diverge in the future
accordingly.